move css styles of debugger-tab to its own file
