+++ src/app/tabs/compile-tab.js
@@ -1,6 +1,5 @@
 /* global */
 var $ = require('jquery')
-
 var yo = require('yo-yo')
 
 var parseContracts = require('../contract/contractParser')
@@ -9,9 +8,9 @@ var modalDialog = require('../ui/modaldialog')
 var modalDialogCustom = require('../ui/modal-dialog-custom')
 var TreeView = require('remix-debugger').ui.TreeView
 var copyToClipboard = require('../ui/copy-to-clipboard')
-var remixLib = require('remix-lib')
 
 var css = require('./styles/compile-tab-styles')
+var remixLib = require('remix-lib')
 var styleGuide = remixLib.ui.themeChooser
 var styles = styleGuide.chooser()
