+++ src/app/tabs/debugger-tab.js
@@ -1,20 +1,5 @@
 var yo = require('yo-yo')
-// -------------- styling ----------------------
-var csjs = require('csjs-inject')
-var styleGuide = require('remix-lib').ui.themeChooser
-var styles = styleGuide.chooser()
-
-var css = csjs`
-  .debuggerTabView {
-    padding: 2%;
-  }
-  .debugger {
-    margin-bottom: 1%;
-    ${styles.rightPanel.debuggerTab.box_Debugger}
-  }
-`
-
-module.exports = debuggerTab
+var css = require('./styles/debugger-tab-styles')
 
 function debuggerTab (container, appAPI, events, opts) {
   var el = yo`
@@ -24,3 +9,5 @@ function debuggerTab (container, appAPI, events, opts) {
   container.appendChild(el)
   return el
 }
+
+module.exports = debuggerTab
